--- src/main/java/com/javarush/ivanenko/MenuWork.java
@@ -1,51 +1,0 @@
-package com.javarush.ivanenko;
-
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
-import java.util.*;
-
-public class MenuWork {
-    public static void Menu(Path path, Scanner scanner, Path resultPath) {
-        System.out.println(Messages.MENU + "\n");
-        while (true) {
-            System.out.print("\n" + Messages.MENU_PROMPT + Messages.ARROW);
-            int num = scanner.nextInt();
-
-            Switch(num, scanner, path, resultPath);
-        }
-    }
-
-    public static void Switch(int num, Scanner scanner, Path path, Path resultPath) {
-        int key;
-        switch (num) {
-            case (1):
-                System.out.print(Messages.ENCRYPTION_KEY_PROMPT + Messages.ARROW);
-                key = scanner.nextInt();
-                Encrypt.encrypt(key, path, resultPath);
-                break;
-            case (2):
-                System.out.print(Messages.DECRYPTION_KEY_PROMPT + Messages.ARROW);
-                key = scanner.nextInt();
-                Decipher.decipher(key, path, resultPath);
-                break;
-            case (3):
-                BruteForceSearch.bruteForce(path, resultPath);
-                break;
-            case (0):
-                System.exit(0);
-                break;
-            default:
-                System.out.println("Введен не верный номер, попробуй ещё раз");
-        }
-    }
-
-    public static char[] keyAlphabet(int key) {
-        char[] keyAlphabet = new char[Messages.ALPHABET.length];
-        for (int i = 0; i < Messages.ALPHABET.length; i++) {
-            keyAlphabet[i] = Messages.ALPHABET[(i + key) % Messages.ALPHABET.length];
-        }
-        return keyAlphabet;
-    }
-}
