+++ src/main/java/com/javarush/khmelov/ConsoleRunner.java
@@ -9,12 +9,12 @@ import java.util.Scanner;
 public class ConsoleRunner {
 
     public static void main(String[] args) {
-        //build console app
+        //build console com.javarush.ivanenko.app
         Scanner input = new Scanner(System.in);
         Menu menu = new Menu(input);
         MainController mainController = new MainController();
         ConsoleApp application = new ConsoleApp(mainController, menu);
-        //run console app
+        //run console com.javarush.ivanenko.app
         application.run(args);
     }
