+++ src/main/java/com/javarush/khmelov/SwingRunner.java
@@ -5,11 +5,11 @@ import com.javarush.khmelov.view.swing.MainForm;
 
 public class SwingRunner {
     public static void main(String[] args) {
-        //build swing app
+        //build swing com.javarush.ivanenko.app
         MainController controller = new MainController();
         MainForm mainForm = new MainForm(controller);
 
-        //run swing app
+        //run swing com.javarush.ivanenko.app
         mainForm.initialization();
     }
 }
